pyproject.toml
[project]
name = "book-scan-to-md"
version = "0.1.0"
description = "Convert book scans (PDF) to markdown using Gemini 2.5 Flash multimodal AI"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
